Working model of NLP score predictor built.
Model used SGDClassifier with input data as CountVectorized and TF-IDF tranfsormed text data. Model is 28% accurate is assigning correct score from 80-100 to a wine, with mean error of 1.6.

Also contains a function desc_predict() that will take in a user string input and output a predicted wine score from the text input data.